requirements.txt
fastapi
motor
python-dotenv
uvicorn
pydantic